+++ src/compas_timber/fabrication/free_contour.py
@@ -1,8 +1,16 @@
+import math
+
 from compas.geometry import Brep
 from compas.geometry import Frame
 from compas.geometry import NurbsCurve
+from compas.geometry import Plane
 from compas.geometry import Polyline
 from compas.geometry import Transformation
+from compas.geometry import Vector
+from compas.geometry import angle_vectors
+from compas.geometry import angle_vectors_signed
+from compas.geometry import distance_point_plane
+from compas.tolerance import Tolerance
 
 from compas_timber.utils import correct_polyline_direction
 from compas_timber.utils import is_polyline_clockwise
@@ -11,6 +19,9 @@ from .btlx import AlignmentType
 from .btlx import BTLxProcessing
 from .btlx import BTLxProcessingParams
 from .btlx import Contour
+from .btlx import DualContour
+
+TOL = Tolerance()
 
 
 class FreeContour(BTLxProcessing):
@@ -18,33 +29,27 @@ class FreeContour(BTLxProcessing):
 
     Parameters
     ----------
-    contour_points : list of :class:`compas.geometry.Point`
-        The points of the contour.
-    depth : float
-        The depth of the contour.
+    contour_param_object : :class:`compas_timber.fabrication.btlx.Contour` or :class:`compas_timber.fabrication.btlx.DualContour`
+        The contour parameter object.
     counter_sink : bool, optional
         If True, the contour is a counter sink. Default is False.
     tool_position : str, optional
         The position of the tool. Default is "left".
     depth_bounded : bool, optional
         If True, the depth is bounded. Default is False, meaning the machining will cut all the way through the element.
-    inclination : float, optional
-        The inclination of the contour. Default is 0. This is not yet implemented.
+    inclination : float or list of float, optional
+        The inclination of the contour. Default is 0 meaning the cut is perpendicular to the reference side.
 
     """
 
     PROCESSING_NAME = "FreeContour"  # type: ignore
 
-    def __init__(self, contour_points, depth, counter_sink=False, tool_position=AlignmentType.LEFT, depth_bounded=False, inclination=0, **kwargs):
+    def __init__(self, contour_param_object, counter_sink=False, tool_position=AlignmentType.LEFT, depth_bounded=True, **kwargs):
         super(FreeContour, self).__init__(**kwargs)
-        self.contour_points = contour_points
-        self.depth = depth
+        self.contour_param_object = contour_param_object
         self.counter_sink = counter_sink
         self.tool_position = tool_position
         self.depth_bounded = depth_bounded
-        if inclination != 0:
-            raise NotImplementedError("Inclination is not yet implemented.")
-        self.inclination = inclination
 
     ########################################################################
     # Properties
@@ -53,12 +58,10 @@ class FreeContour(BTLxProcessing):
     @property
     def __data__(self):
         data = super(FreeContour, self).__data__
-        data["contour_points"] = self.contour_points
-        data["depth"] = self.depth
+        data["contour_param_object"] = self.contour_param_object
         data["counter_sink"] = self.counter_sink
         data["tool_position"] = self.tool_position
         data["depth_bounded"] = self.depth_bounded
-        data["inclination"] = self.inclination
         return data
 
     @property
@@ -70,7 +73,7 @@ class FreeContour(BTLxProcessing):
     ########################################################################
 
     @classmethod
-    def from_polyline_and_element(cls, polyline, element, depth=None, interior=True, tool_position=None, ref_side_index=3):
+    def from_polyline_and_element(cls, polyline, element, depth=None, interior=None, tool_position=None, ref_side_index=None):
         """Construct a Contour processing from a polyline and element.
 
         Parameters
@@ -80,37 +83,75 @@ class FreeContour(BTLxProcessing):
         element : :class:`compas_timber.elements.Beam` or :class:`compas_timber.elements.Plate`
             The element.
         depth : float, optional
-            The depth of the contour. Default is the width of the element.
+            The depth of the contour. Default is the thickness of the element.
         interior : bool, optional
             If True, the contour is an interior contour. Default is True.
-        tool_position : str, optional
+        tool_position : BTLx.AlignmentType, optional
             The position of the tool. Default is "left".
         ref_side_index : int, optional
-            The reference side index. Default is 3.
+            The reference side index. If none is given, the function will try to find the reference side index based on the polyline and element.
         """
-        if polyline[0] != polyline[-1]:  # if the polyline is not closed
-            if not tool_position:
-                raise ValueError("The polyline should be closed or a tool position should be provided.")
-            else:
-                return cls(polyline, depth, tool_position=tool_position, ref_side_index=ref_side_index)
 
-        if interior:
-            counter_sink = True
-            if is_polyline_clockwise(polyline, element.ref_sides[ref_side_index].normal):
-                tool_position = AlignmentType.RIGHT
-            else:
-                tool_position = AlignmentType.LEFT
-        else:
-            counter_sink = False
-            if is_polyline_clockwise(polyline, element.ref_sides[ref_side_index].normal):
-                tool_position = AlignmentType.LEFT
-            else:
-                tool_position = AlignmentType.RIGHT
+        if ref_side_index is None:
+            ref_side_index = cls.get_ref_face_index(polyline, element)
 
+        ref_side = element.ref_sides[ref_side_index]
+        tool_position = cls.parse_tool_position(polyline, ref_side, tool_position, interior)
         depth = depth or element.width
-        xform = Transformation.from_frame_to_frame(element.ref_frame, Frame.worldXY())
-        points = [pt.transformed(xform) for pt in polyline]
-        return cls(points, depth, tool_position=tool_position, counter_sink=counter_sink, ref_side_index=ref_side_index)
+        xform_to_part_coords = Transformation.from_frame_to_frame(ref_side, Frame.worldXY())
+        points = [pt.transformed(xform_to_part_coords) for pt in polyline]
+        contour = Contour(points, depth=depth, inclination=[0.0])
+
+        return cls(contour, tool_position=tool_position, counter_sink=interior, ref_side_index=ref_side_index)
+
+    @classmethod
+    def from_polylines_and_element(cls, polylines, element, interior=None, tool_position=None, ref_side_index=None):
+        """Construct a Contour processing from a list of polylines and element.
+
+        Parameters
+        ----------
+        polylines : list of list of :class:`compas.geometry.Point`
+            The top and bottome polylines of the contour.
+        element : :class:`compas_timber.elements.Beam` or :class:`compas_timber.elements.Plate`
+            The element.
+        interior : bool, optional
+            If True, the contour is an interior contour. Default is None.
+        tool_position : BTLx.AlignmentType, optional
+            The position of the tool. Default is None.
+        ref_side_index : int, optional
+            The reference side index. If none is given, the function will try to find the reference side index based on the polylines and element.
+        """
+
+        if len(polylines) != 2:
+            raise ValueError("Two polylines are expected.")
+
+        if len(polylines[0]) != len(polylines[1]):
+            raise ValueError("The top and bottom polylines should have the same number of points.")
+
+        if not ref_side_index:
+            ref_side_index = cls.get_ref_face_index(polylines[0], element)
+
+        ref_side = element.ref_sides[ref_side_index]
+        xform_to_part_coords = Transformation.from_frame_to_frame(ref_side, Frame.worldXY())
+        tool_position = cls.parse_tool_position(polylines[0], ref_side, tool_position, interior)
+
+        if not cls.are_all_segments_parallel(polylines[0], polylines[1]):  # use DualContour
+            points_principal = [pt.transformed(xform_to_part_coords) for pt in polylines[0]]
+            points_associated = [pt.transformed(xform_to_part_coords) for pt in polylines[1]]
+            dual_contour = DualContour(points_principal, points_associated)
+            return cls(dual_contour, counter_sink=interior, tool_position=tool_position, ref_side_index=ref_side_index)
+
+        else:  # use Contour with inclination
+            inclinations = []
+            for top_line, bottom_line in zip(Polyline(polylines[0]).lines, Polyline(polylines[1]).lines):
+                cp = bottom_line.closest_point(top_line.start)
+                inclinations.append(round(angle_vectors_signed(Vector.from_start_end(top_line.start, cp), -ref_side.normal, -top_line.direction, deg=True), 6))
+            if len(set(inclinations)) == 1:
+                inclinations = [inclinations[0]]  # all inclinations are the same, set one global inclination for FreeContour processing
+            depth = distance_point_plane(polylines[1][0], Plane.from_frame(ref_side))
+            polyline = Polyline(polylines[0]).transformed(xform_to_part_coords)
+            contour = Contour(polyline, depth=depth, inclination=inclinations)
+            return cls(contour, counter_sink=interior, tool_position=tool_position, ref_side_index=ref_side_index)
 
     @classmethod
     def from_shapes_and_element(cls, polyline, element, depth=None, interior=True, **kwargs):
@@ -129,6 +170,48 @@ class FreeContour(BTLxProcessing):
         """
         return cls.from_polyline_and_element(polyline, element, depth, interior, **kwargs)
 
+    @staticmethod
+    def parse_tool_position(polyline, ref_side, tool_position, interior):
+        if polyline[0] != polyline[-1]:  # if polyline is not closed
+            if tool_position is None:
+                raise ValueError("The polyline should be closed or a tool position should be provided.")
+            elif interior:
+                raise ValueError("Interior polyline must be closed.")
+            else:
+                return tool_position
+        else:  # if the polyline is closed
+            if interior:
+                if is_polyline_clockwise(polyline, ref_side.normal):
+                    calculated_tool_position = AlignmentType.RIGHT
+                else:
+                    calculated_tool_position = AlignmentType.LEFT
+            else:
+                if is_polyline_clockwise(polyline, ref_side.normal):
+                    calculated_tool_position = AlignmentType.LEFT
+                else:
+                    calculated_tool_position = AlignmentType.RIGHT
+            if tool_position is None or tool_position == calculated_tool_position:
+                return calculated_tool_position
+            else:
+                raise ValueError("Tool position does not match the contour direction for internal contour.")
+
+    @staticmethod
+    def get_ref_face_index(contour_points, element):
+        curve_frame = Frame.from_points(contour_points[0], contour_points[1], contour_points[-2])
+        for i, ref_side in enumerate(element.ref_sides):
+            if TOL.is_zero(distance_point_plane(contour_points[0], Plane.from_frame(ref_side)), tol=1e-6) and TOL.is_zero(
+                angle_vectors(ref_side.normal, curve_frame.zaxis, deg=True) % 180.0, 1e-6
+            ):
+                return i
+        raise ValueError("The contour does not lay on one of the reference sides of the element.")
+
+    @staticmethod
+    def are_all_segments_parallel(polyline_a, polyline_b):
+        for top_line, bottom_line in zip(Polyline(polyline_a).lines, Polyline(polyline_b).lines):
+            if not TOL.is_zero(angle_vectors(top_line.direction, bottom_line.direction) % math.pi, tol=1e-6):
+                return False
+        return True
+
     ########################################################################
     # Methods
     ########################################################################
@@ -146,11 +229,12 @@ class FreeContour(BTLxProcessing):
         :class:`compas.geometry.Brep`
             The resulting geometry after processing.
         """
-        xform = Transformation.from_frame_to_frame(Frame.worldXY(), element.ref_sides[self.ref_side_index])
+        ref_side = element.ref_sides[self.ref_side_index]
+        xform = Transformation.from_frame_to_frame(Frame.worldXY(), ref_side)
         pts = [pt.transformed(xform) for pt in self.contour_points]
-        pts = correct_polyline_direction(pts, element.ref_sides[self.ref_side_index].normal, clockwise=True)
-        vol = Brep.from_extrusion(NurbsCurve.from_points(pts, degree=1), element.ref_sides[self.ref_side_index].normal * self.depth * 2.0)
-        vol.translate(element.ref_sides[self.ref_side_index].normal * -self.depth)
+        pts = correct_polyline_direction(pts, ref_side.normal, clockwise=True)
+        vol = Brep.from_extrusion(NurbsCurve.from_points(pts, degree=1), ref_side.normal * self.depth * 2.0)
+        vol.translate(ref_side.normal * -self.depth)
 
         if self.counter_sink:  # contour should remove material inside of the contour
             return geometry - vol
@@ -164,8 +248,7 @@ class FreeCountourParams(BTLxProcessingParams):
         super(FreeCountourParams, self).__init__(instance)
 
     def as_dict(self):
-        contour_line = Polyline(self._instance.contour_points)
-        return {"Contour": Contour(contour_line, depth=self._instance.depth, depth_bounded=self._instance.depth_bounded, inclination=self._instance.inclination)}
+        return {"Contour": self._instance.contour_param_object}
 
     @property
     def header_attributes(self):
@@ -176,5 +259,5 @@ class FreeCountourParams(BTLxProcessingParams):
             "ToolID": "0",
             "Process": "yes",
             "ToolPosition": self._instance.tool_position,
-            "ReferencePlaneID": "4",
+            "ReferencePlaneID": str(self._instance.ref_side_index + 1),
         }